requirements.txt
fastapi
uvicorn
google-analytics-data
openai
python-dateutil
python-dotenv
pandas